pyproject.toml
[tool.poetry]
name = "SHIFT_Test_Task"
version = "0.1.0"
description = ""
authors = ["angheelina <ahmetovaanghelina@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
bcrypt = "4.0.1"
sqlalchemy = "^2.0.30"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
pydantic-settings = "^2.2.1"
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
